+++ packages/camera/camera_android/android/src/main/java/io/flutter/plugins/camera/Camera.java
@@ -107,8 +107,10 @@ class Camera
   private CameraProperties cameraProperties;
   private final CameraFeatureFactory cameraFeatureFactory;
   private final Activity activity;
+
   /** A {@link CameraCaptureSession.CaptureCallback} that handles events related to JPEG capture. */
   private final CameraCaptureCallback cameraCaptureCallback;
+
   /** A {@link Handler} for running tasks in the background. */
   Handler backgroundHandler;
 
@@ -119,12 +121,15 @@ class Camera
   CameraCaptureSession captureSession;
   @VisibleForTesting ImageReader pictureImageReader;
   ImageStreamReader imageStreamReader;
+
   /** {@link CaptureRequest.Builder} for the camera preview */
   CaptureRequest.Builder previewRequestBuilder;
 
   @VisibleForTesting MediaRecorder mediaRecorder;
+
   /** True when recording video. */
   boolean recordingVideo;
+
   /** True when the preview is paused. */
   @VisibleForTesting boolean pausedPreview;
 
@@ -132,6 +137,7 @@ class Camera
 
   /** Holds the current capture timeouts */
   private CaptureTimeoutsWrapper captureTimeouts;
+
   /** Holds the last known capture properties */
   private CameraCaptureProperties captureProps;
 
@@ -271,8 +277,9 @@ class Camera
 
     MediaRecorderBuilder mediaRecorderBuilder;
 
-    // TODO(camsim99): Revert changes that allow legacy code to be used when recordingProfile is null
-    // once this has largely been fixed on the Android side. https://github.com/flutter/flutter/issues/119668
+    // TODO(camsim99): Revert changes that allow legacy code to be used when recordingProfile
+    // is null once this has largely been fixed on the Android side.
+    // https://github.com/flutter/flutter/issues/119668
     if (SdkCapabilityChecker.supportsEncoderProfiles() && getRecordingProfile() != null) {
       mediaRecorderBuilder =
           new MediaRecorderBuilder(
@@ -1157,7 +1164,8 @@ class Camera
 
   public void startPreview(@Nullable Runnable onSuccessCallback)
       throws CameraAccessException, InterruptedException {
-    // If recording is already in progress, the camera is being flipped, so send it through the VideoRenderer to keep the correct orientation.
+    // If recording is already in progress, the camera is being flipped, so send it through the
+    // VideoRenderer to keep the correct orientation.
     if (recordingVideo) {
       startPreviewWithVideoRendererStream(onSuccessCallback);
     } else {
@@ -1379,7 +1387,8 @@ class Camera
           "setDescriptionWhileRecordingFailed", "Device was not recording", null);
     }
 
-    // See VideoRenderer.java; support for this EGL extension is required to switch camera while recording.
+    // See VideoRenderer.java; support for this EGL extension is required to switch camera while
+    // recording.
     if (!SdkCapabilityChecker.supportsEglRecordableAndroid()) {
       throw new Messages.FlutterError(
           "setDescriptionWhileRecordingFailed",