+++ packages/image_picker/image_picker/CHANGELOG.md
@@ -1,10 +1,12 @@
-## NEXT
+## 1.2.2
 
+* Clarifies that `maxDuration` only applies to video recording and not gallery selection.
 * Updates minimum supported SDK version to Flutter 3.35/Dart 3.9.
 * Updates README to reflect currently supported OS versions for the latest
   versions of the endorsed platform implementations.
   * Applications built with older versions of Flutter will continue to
     use compatible versions of the platform implementations.
+
 
 ## 1.2.1